requirements.txt
# ============================================
# DYLUMO - Image to Music Recommendation
# Requirements for Training & Inference
# ============================================

# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Image Feature Extraction
transformers>=4.35.0
Pillow>=10.0.0
open-clip-torch>=2.23.0
timm>=0.9.0

# Similarity Search
faiss-cpu>=1.7.4

# Data Processing & Utilities
tqdm>=4.66.0
requests>=2.31.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
regex>=2023.0.0
ftfy>=6.1.0

# HuggingFace
huggingface_hub>=0.19.0
datasets>=2.14.0
safetensors>=0.4.0
tokenizers>=0.15.0

# Kaggle API
kaggle>=1.5.16

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter support
ipykernel>=6.25.0
ipython>=8.0.0

# Additional utilities
joblib>=1.3.0
PyYAML>=6.0.0
Jinja2>=3.1.0
